move and update tests
