+++ tests/Fixtures/inc/Engine/Media/ImagesDimensions/ImagesDimensions/addOption.php
@@ -1,23 +1,26 @@
 <?php
 
 return [
-	'shouldReturnValidOptionsWithEmptyOptions' => [
-		'input' => [
-			'options' => [],
-		],
-		'expected' => [
-			'images_dimensions' => 0,
-		]
-	],
-	'shouldNotOverrideOtherOptions' => [
-		'input' => [
-			'options' => [
-				'test_option' => 1,
+	'vfs_dir'   => '/',
+	'test_data' => [
+		'shouldReturnValidOptionsWithEmptyOptions' => [
+			'input' => [
+				'options' => [],
 			],
+			'expected' => [
+				'images_dimensions' => 0,
+			]
 		],
-		'expected' => [
-			'test_option'       => 1,
-			'images_dimensions' => 0,
-		]
+		'shouldNotOverrideOtherOptions' => [
+			'input' => [
+				'options' => [
+					'test_option' => 1,
+				],
+			],
+			'expected' => [
+				'test_option'       => 1,
+				'images_dimensions' => 0,
+			]
+		],
 	],
 ];
