+++ tests/Fixtures/inc/Engine/Media/ImagesDimensions/empty.jpg
